+++ test/libsolidity/smtCheckerTests/types/tuple_1_chain_n.sol
@@ -1,0 +1,10 @@
+pragma experimental SMTChecker;
+contract C {
+	function i() public pure returns (uint d) {
+		if (0==0)
+			(((((d))))) = 13;
+		assert(d == 13);
+	}
+}
+// ----
+// Warning 6838: (96-100): Condition is always true.